refactor: login and signup page
