+++ src/scss/_global.scss
@@ -17,7 +17,7 @@ body {
   width: 100%;
   min-height: 100vh;
   overflow-x: hidden;
-  overflow-y: scroll;
+  overflow-y: auto;
 }
 
 .grid {
@@ -26,7 +26,7 @@ body {
   margin: 0 auto;
 }
 
-.btn {
+button {
   font-size: 1.6rem;
   font-weight: 600;
   line-height: 2.4rem;
@@ -36,14 +36,14 @@ body {
 
   border: none;
   border-radius: 4px;
-
   cursor: pointer;
 
   transition-duration: 0.15s, 0.15s, 0.15s, 0.15s;
   transition-property: color, background-color, border-color, box-shadow;
   transition-timing-function: ease-in-out, ease-in-out, ease-in-out, ease-in-out;
+}
 
-  // box-shadow: $btn-box-shadow;
+.btn {
   @include button-mixin(default, $primary-color);
 
   @each $variant in $btn-variants {
