+++ src/scss/authentication.scss
@@ -1,0 +1,84 @@
+@import "./_mixins";
+@import "./variables";
+
+.auth {
+  &_container {
+    width: 100%;
+    min-height: calc(100vh - 80px);
+    margin: 0 auto;
+    padding: 120px 32px;
+
+    display: flex;
+    justify-content: center;
+    align-items: flex-start;
+
+    background: $bg-gradient-color;
+  }
+
+  &_banner {
+    flex: 1;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  &_content {
+    flex: 1;
+  }
+
+  &_form {
+    &-input {
+      width: 400px;
+      font-size: 1.6rem;
+
+      margin-top: 16px;
+      padding: 18px 12px;
+
+      background-color: transparent;
+
+      border: 1px solid $light-grey-color;
+      border-color: $secondary;
+      border-radius: 6px;
+      outline: $primary-color;
+
+      transition: all 0.3s ease-in-out;
+      box-shadow: rgba(182, 182, 182, 0.18) 0px 4px 10px 0px;
+
+      &:focus,
+      &:hover {
+        border: 1px solid $black-color;
+      }
+
+      &--error {
+        border: 1px solid $red-color;
+
+        &:focus,
+        &:hover {
+          border: 1px solid $red-color;
+        }
+      }
+    }
+
+    &-message {
+      width: 100%;
+      font-size: 1.3rem;
+      font-weight: 500;
+
+      &--error {
+        color: $red-color;
+      }
+    }
+  }
+
+  &_submit-btn {
+    @include button-mixin("default", $primary);
+    width: 100%;
+    margin-top: 16px;
+    padding: 12px 24px;
+    font-size: 1.8rem;
+
+    &--loading {
+      @include button-mixin("default", $secondary);
+    }
+  }
+}
